Controller에서 deleteTask 메소드 수정하기

+++ calendar/src/main/java/com/study/calendar/repository/TaskRepository.java
@@ -12,5 +12,5 @@ import java.util.List;
 public interface TaskRepository extends JpaRepository<Task, Long> {
     List<Task> findByUser(User user);
     List<Task> findByDeadlineDate(LocalDate deadlineDate);
-   void deleteById(Integer id);
+   void deleteById(Long id);
 }
